--- src/js/pages/FallbackArbitration/index.scss
@@ -1,47 +1,0 @@
-@import "../../../css/variable-overrides";
-
-.arbitrationStatus {
-  border-radius: 4px;
-  background: $black;
-  text-transform: uppercase;
-  color: $white;
-  font-weight: bold;
-  font-size: 11px;
-  letter-spacing: 1px;
-  padding: 3px 10px;
-  position: relative;
-  top: -3px;
-}
-
-.arbitrationDialog button svg {
-  margin-top: 10px !important;
-}
-
-.arbitrationStatus.open {
-  background: $success;
-}
-
-.arbitrationStatus.resolved {
-  background: $primary;
-}
-
-.arbitrationMotive {
-  background: $secondary;
-  border-radius: 8px;
-  padding: 10px;
-}
-
-.triangle {
-  display: block;
-  position: relative;
-  top: -15px;
-  left: 18px;
-  margin-bottom: -8px;
-}
-
-body .seller-text {
-  background-color: $gray-300;
-}
-body .buyer-text {
-  background-color: $gray-100;
-}
